+++ rs/proposals/src/canisters/mod.rs
@@ -2,3 +2,4 @@
 pub mod nns_governance;
 pub mod nns_registry;
 pub mod sns_wasm;
+pub mod internet_identity;
